django/requirements/base.txt
certifi==2016.2.28
coverage==3.7
Django==1.9.6
django-braces==1.8.1
django-rest-swagger==0.3.7
djangorestframework==3.3.3
django-cors-headers==1.1.0
mock==1.0.1
mysqlclient==1.3.7
oauthlib==1.0.3
PyYAML==3.11
six==1.10.0
python-logstash==0.4.6
dj-static==0.0.6
gevent==1.1.2